pre_training/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
allennlp==0.9.0
asn1crypto==0.24.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
aws-sam-translator==1.11.0
aws-xray-sdk==2.4.2
awscli==1.16.161
Babel==2.7.0
backcall==0.2.0
bleach==3.1.5
blis==0.2.4
boto==2.49.0
boto3==1.9.228
botocore==1.12.228
brotlipy==0.7.0
cachetools==4.0.0
certifi==2020.4.5.2
cffi==1.14.0
cfn-lint==0.20.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
conllu==1.3.1
cryptography==2.9.2
cxxfilt==0.2.0
cycler==0.10.0
cymem==2.0.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docker==4.0.1
docutils==0.15.2
ecdsa==0.13.2
editdistance==0.5.3
en-core-web-sm==2.1.0
entrypoints==0.3
filelock==3.0.12
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.17.1
gast==0.2.2
gdown==3.11.0
gevent==1.4.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
googleapis-common-protos==1.6.0
GPUtil==1.4.0
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
idna==2.9
imagesize==1.1.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1592348411427/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.13.2
json5==0.9.5
jsondiff==1.1.2
jsonlines==1.2.0
jsonnet==0.13.0
jsonpatch==1.23
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.0.0
moto==1.3.8
msgpack==0.5.6
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
overrides==1.9
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.9.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.1.2
python-dateutil==2.8.1
python-jose==3.0.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.1
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.10.6
rsa==3.4.2
s3transfer==0.2.1
sacremoses==0.0.35
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.15.0
snowballstemmer==1.9.1
spacy==2.1.8
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.1.0
tb-nightly==1.14.0a20190520
tensorboard==2.0.2
tensorboardX==1.8
tensorflow==2.0.0
tensorflow-datasets==1.3.2
tensorflow-estimator==2.0.1
tensorflow-metadata==0.15.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.0.8
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tornado==6.0.4
tqdm==4.35.0
traitlets==4.3.3
transformers==2.3.0
ujson==1.35
Unidecode==1.1.1
urllib3==1.25.9
wasabi==0.2.2
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.6
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.11.1
xmltodict==0.12.0
zipp==3.1.0
